LPS-135551 Rename to *Impl
